+++ tests/Pipaslot.Mediator.Http.Tests/MediatorMiddlewareTests.cs
@@ -150,120 +150,6 @@ public class MediatorMiddlewareTests
         Assert.False(second.Applied);
     }
 
-    [Fact]
-    public async Task WillApplyStatusCodeHint_WhenPresent()
-    {
-        var mediatorResponse = Task.FromResult((IMediatorResponse)new MediatorResponse(true, [new ResponseStatusCodeHint(400)]));
-        var mediatorMock = new Mock<IMediator>();
-        mediatorMock.Setup(x => x.Dispatch(It.IsAny<IMediatorAction>(), It.IsAny<CancellationToken>())).Returns(mediatorResponse);
-        var services = CreateServiceProvider(mediatorMock);
-        var sut = services.GetRequiredService<MediatorMiddleware>();
-
-        var response = new FakeResponse();
-        var context = new FakeContext(new FakePostRequest(_message), services, response);
-        await sut.Invoke(context);
-
-        Assert.Equal(400, response.StatusCode);
-    }
-
-    [Fact]
-    public async Task WillOverwriteAlreadySetStatusCode_WhenHintPresent()
-    {
-        // A hint always wins over whatever a legacy middleware already wrote directly to HttpContext.Response.
-        var mediatorResponse = Task.FromResult((IMediatorResponse)new MediatorResponse(false, [new ResponseStatusCodeHint(409)]));
-        var mediatorMock = new Mock<IMediator>();
-        mediatorMock.Setup(x => x.Dispatch(It.IsAny<IMediatorAction>(), It.IsAny<CancellationToken>())).Returns(mediatorResponse);
-        var services = CreateServiceProvider(mediatorMock);
-        var sut = services.GetRequiredService<MediatorMiddleware>();
-
-        var response = new FakeResponse { StatusCode = (int)HttpStatusCode.BadRequest };
-        var context = new FakeContext(new FakePostRequest(_message), services, response);
-        await sut.Invoke(context);
-
-        Assert.Equal(409, response.StatusCode);
-    }
-
-    [Fact]
-    public async Task WillUseLastStatusCodeHint_WhenMultiplePresent()
-    {
-        // A later hint overrides an earlier one, mirroring what MediatorContext.Features.Set() would do.
-        var mediatorResponse = Task.FromResult((IMediatorResponse)new MediatorResponse(true,
-            [new ResponseStatusCodeHint(400), new ResponseStatusCodeHint(409)]));
-        var mediatorMock = new Mock<IMediator>();
-        mediatorMock.Setup(x => x.Dispatch(It.IsAny<IMediatorAction>(), It.IsAny<CancellationToken>())).Returns(mediatorResponse);
-        var services = CreateServiceProvider(mediatorMock);
-        var sut = services.GetRequiredService<MediatorMiddleware>();
-
-        var response = new FakeResponse();
-        var context = new FakeContext(new FakePostRequest(_message), services, response);
-        await sut.Invoke(context);
-
-        Assert.Equal(409, response.StatusCode);
-    }
-
-    [Fact]
-    public async Task WillNotApplyStatusCodeHint_WhenResponseAlreadyStarted()
-    {
-        var mediatorResponse = Task.FromResult((IMediatorResponse)new MediatorResponse(true, [new ResponseStatusCodeHint(400)]));
-        var mediatorMock = new Mock<IMediator>();
-        mediatorMock.Setup(x => x.Dispatch(It.IsAny<IMediatorAction>(), It.IsAny<CancellationToken>())).Returns(mediatorResponse);
-        var services = CreateServiceProvider(mediatorMock);
-        var sut = services.GetRequiredService<MediatorMiddleware>();
-
-        var response = new FakeResponse(hasStarted: true);
-        var context = new FakeContext(new FakePostRequest(_message), services, response);
-        await sut.Invoke(context);
-
-        Assert.Equal((int)HttpStatusCode.OK, response.StatusCode);
-        // Default JSON serialization must also be skipped once something else already wrote to the response
-        Assert.Equal(string.Empty, response.ContentType);
-    }
-
-    [Fact]
-    public async Task WillIgnoreStatusCodeHint_WhenHttpResultPresent()
-    {
-        // IMediatorHttpResult owns the whole response and takes precedence over a mere status-code annotation.
-        var httpResult = new FakeHttpResult();
-        var mediatorResponse = Task.FromResult((IMediatorResponse)new MediatorResponse(true, [httpResult, new ResponseStatusCodeHint(400)]));
-        var mediatorMock = new Mock<IMediator>();
-        mediatorMock.Setup(x => x.Dispatch(It.IsAny<IMediatorAction>(), It.IsAny<CancellationToken>())).Returns(mediatorResponse);
-        var services = CreateServiceProvider(mediatorMock);
-        var sut = services.GetRequiredService<MediatorMiddleware>();
-
-        var response = new FakeResponse();
-        var context = new FakeContext(new FakePostRequest(_message), services, response);
-        await sut.Invoke(context);
-
-        Assert.True(httpResult.Applied);
-        Assert.Equal((int)HttpStatusCode.OK, response.StatusCode);
-    }
-
-    [Fact]
-    public async Task WillExcludeStatusCodeHint_FromSerializedResponseBody()
-    {
-        const string otherResult = "actual-result";
-        var mediatorResponse = Task.FromResult((IMediatorResponse)new MediatorResponse(true, [otherResult, new ResponseStatusCodeHint(400)]));
-        var mediatorMock = new Mock<IMediator>();
-        mediatorMock.Setup(x => x.Dispatch(It.IsAny<IMediatorAction>(), It.IsAny<CancellationToken>())).Returns(mediatorResponse);
-
-        IMediatorResponse? serialized = null;
-        var serializerMock = new Mock<IContractSerializer>();
-        serializerMock.Setup(x => x.SerializeResponse(It.IsAny<IMediatorResponse>()))
-            .Callback<IMediatorResponse>(r => serialized = r)
-            .Returns("{}");
-
-        var services = CreateServiceProvider(mediatorMock, serializerMock);
-        var sut = services.GetRequiredService<MediatorMiddleware>();
-
-        var response = new FakeResponse();
-        var context = new FakeContext(new FakePostRequest(_message), services, response);
-        await sut.Invoke(context);
-
-        Assert.NotNull(serialized);
-        Assert.DoesNotContain(serialized!.Results, r => r is ResponseStatusCodeHint);
-        Assert.Contains(otherResult, serialized.Results);
-    }
-
     /// <summary>
     /// Integration-level sanity check for safe-by-default error handling across the real HTTP pipeline: a real
     /// handler throws, the dispatch boundary converts it to a generic failure, and the JSON serializer writes the
@@ -339,7 +225,7 @@ public class MediatorMiddlewareTests
         mediatorMock.Verify(m => m.Dispatch(It.IsAny<NopMessage>(), It.IsAny<CancellationToken>()));
     }
 
-    private ServiceProvider CreateServiceProvider(Mock<IMediator> mediatorMock, Mock<IContractSerializer>? serializerMock = null)
+    private ServiceProvider CreateServiceProvider(Mock<IMediator> mediatorMock)
     {
         var collection = new ServiceCollection();
         collection.AddLogging();
@@ -348,10 +234,6 @@ public class MediatorMiddlewareTests
         collection.AddScoped<MediatorMiddleware>();
         collection.AddScoped<RequestDelegate>(s => (c) => Task.CompletedTask);
         collection.AddSingleton<IMediator>(mediatorMock.Object);
-        if (serializerMock is not null)
-        {
-            collection.AddSingleton(serializerMock.Object);
-        }
         return collection.BuildServiceProvider();
     }
 }
